+++ src/main/java/com/github/games647/flexiblelogin/FlexibleLogin.java
@@ -33,7 +33,7 @@ import org.spongepowered.api.text.Texts;
 import org.spongepowered.api.util.command.args.GenericArguments;
 import org.spongepowered.api.util.command.spec.CommandSpec;
 
-@Plugin(id = "flexiblelogin", name = "FlexibleLogin", version = "0.2")
+@Plugin(id = "flexiblelogin", name = "FlexibleLogin", version = "0.2.1")
 public class FlexibleLogin {
 
     private final PluginContainer pluginContainer;